Since not more than one module has been used for the application source code its main one is renamed back to `super_simple_stocks.py`.
