+++ super_simple_stocks/tests/fixture_data.py
@@ -1,7 +1,7 @@
-from ..model import (TickerSymbol,
-                     CommonStock,
-                     PreferredStock,
-                     BuySellIndicator)
+from ..super_simple_stocks import (TickerSymbol,
+                                   CommonStock,
+                                   PreferredStock,
+                                   BuySellIndicator)
 
 STOCKS = (
     (TickerSymbol.TEA, CommonStock, 0.0, None, 100.0),
